Added bios support... (Limited.)

+++ airootfs/install/postinst/bootloader.sh
@@ -6,7 +6,14 @@ mkinitcpio -p linux
 echo "Installing packages..."
 sleep 5
 pacman -S efibootmgr grub os-prober
-echo "Installing GRUB..."
-grub-install --target=x86_64-efi --efi-directory=boot/efi --bootloader-id=BlueOS
+
+FILE=/bios
+if test -f "$FILE"
+then
+	grub-install /dev/sda
+else
+	echo "Installing GRUB..."
+	grub-install --target=x86_64-efi --efi-directory=boot/efi --bootloader-id=BlueOS
+fi
 echo "Generating GRUB config...."
 grub-mkconfig -o /boot/grub/grub.cfg
